Update dependency on oj
I'm using pronto-json as a test suite dependency in one of my Rails applications, but it's causing me to be locked into a really old version of the `oj` gem. `oj` doesn't really break compatibility with their releases, so it should be safe to set a newer dependency (or no specific version dependency at all.